AI-Money-Mentor-main/backend/requirements.txt
yfinance>=1.2.0
pandas>=2.3.0
# pandas-ta>=0.3.14b1  # Optional: requires Python 3.12+ for technical analysis indicators
streamlit>=1.55.0
fastapi>=0.135.0
uvicorn>=0.42.0
plotly>=6.6.0
requests>=2.32.0
mftool>=3.2
casparser>=0.8.1
numpy>=2.2.0
scipy>=1.11.0
python-telegram-bot>=21.0
# openai>=1.0.0  # Not used directly — AI handled by OpenClaw + Ollama/GLM-5 Cloud
python-dotenv
